requirements.txt
python-dotenv
langchain-groq
langchain-huggingface
langchain-core
langchain-community
faiss-cpu
PyPDF2
transformers
torch
pypdf
fastapi
uvicorn
sse-starlette
requests
bs4
fastapi
uvicorn
langchain
langchain-community
faiss-cpu
pydantic

